Anonymize/src/requirements.txt
boto3
pytz
requests
flask
pymongo
dnspython